moved stuff back into common for now
